add expression input and work with double numbers

+++ src/MyEcxeption.java
@@ -1,12 +1,26 @@
-// Our class exception, which our methods throw when find error in expression
+import java.math.BigDecimal;
+
+/**
+ * Our class exception, which our methods throw when find error in expression
+ * implements {@link ExpressionComponents}
+ * @version  1.0
+ * @author  Hakop Hakopyan
+ */
 public class MyEcxeption extends Exception implements  ExpressionComponents{
     private  String myException;
 
+    /**
+     * @return type String, return value {@link MyEcxeption#myException}
+     */
     public String getMyException() {
         return myException;
     }
 
-    // check exception and found the appropriate error description
+    /**
+     * check exception and found the appropriate error description
+     * Added only BRACKETS error description
+     * @param exception contains character
+     */
     private void setMyExceptionWithCheck(String exception) {
         String finalException = "Error: ";
         switch (exception) {
@@ -21,13 +35,20 @@ public class MyEcxeption extends Exception implements  ExpressionComponents{
         setMyException(finalException);
     }
 
-    // Use when there is a description of the error
+    /**
+     * Use when there is a description of the error
+     * @param exception contains description of the error
+     */
     private  void setMyException(String exception) {
             this.myException = "Error: " + exception;
     }
 
-    // Use when needed to found correct error description
-    // Added only BRACKETS error description
+    /**
+     * Use when needed to found correct error description
+     * Added only BRACKETS error description
+     * @param exception contains character or error description
+     * @param check determines search for character the appropriate error or not
+     */
     MyEcxeption(String exception, boolean check) {
         if (check)
             setMyExceptionWithCheck(exception);
@@ -35,12 +56,21 @@ public class MyEcxeption extends Exception implements  ExpressionComponents{
             setMyException(exception);
     }
 
+    /**
+     * Use when there is a ready error message
+     * @param exception contains description of the error
+     */
     MyEcxeption(String exception) {
         setMyException(exception);
     }
 
-    // Use when Error with arithmetic operation
-    MyEcxeption(double leftArg, double rightArg, String operator) {
+    /**
+     * Use when Error with arithmetic operation error
+     * @param leftArg the number from the left of the operator
+     * @param rightArg the number from the right of the operator
+     * @param operator contains operator
+     */
+    MyEcxeption(BigDecimal leftArg, BigDecimal rightArg, String operator) {
         String finalException = "";
         if (operator == DIVISION) {
             finalException += "Division by zero: ";
